+++ Source/GUI/SquidSalmple/CueSets/WaveformDisplay.h
@@ -32,7 +32,7 @@ private:
     juce::int64 numSamples { 0 };
     int halfHeight { 0 };
     int numPixels { 0 };
-    int samplesPerPixel { 0 };
+    float samplesPerPixel { 0.f };
     int markerStartY { 1 };
     int markerEndY { 0 };
     std::array<float, 2> dashedSpec;
